Add tests for CallerFree for strings, and refactor test naming.
