+++ AdvancedDLSupport.Tests/c/CMakeLists.txt
@@ -39,7 +39,7 @@ add_library(FunctionTests SHARED src/FunctionTests.c ${SHARED_HEADERS})
 add_library(LazyLoadingTests SHARED src/LazyLoadingTests.c ${SHARED_HEADERS})
 add_library(PropertyTests SHARED src/PropertyTests.c ${SHARED_HEADERS})
 add_library(RemappingTests SHARED src/RemappingTests.c ${SHARED_HEADERS})
-add_library(TypeLoweringTests SHARED src/TypeLoweringTests.c ${SHARED_HEADERS})
+add_library(StringTests SHARED src/StringTests.c ${SHARED_HEADERS})
 add_library(AttributePassthroughTests SHARED src/AttributePassthroughTests.c ${SHARED_HEADERS})
 add_library(MixedModeTests SHARED src/MixedModeTests.c ${SHARED_HEADERS})
 add_library(NullableTests SHARED src/NullableTests.c ${SHARED_HEADERS})
@@ -56,7 +56,7 @@ if (IS_UNIX_COMPILER)
         set_target_properties(LazyLoadingTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
         set_target_properties(PropertyTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
         set_target_properties(RemappingTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
-        set_target_properties(TypeLoweringTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
+        set_target_properties(StringTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
         set_target_properties(AttributePassthroughTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
         set_target_properties(MixedModeTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
         set_target_properties(NullableTests PROPERTIES COMPILE_FLAGS "-m64" LINK_FLAGS "-m64")
@@ -70,7 +70,7 @@ if (IS_UNIX_COMPILER)
         set_target_properties(LazyLoadingTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
         set_target_properties(PropertyTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
         set_target_properties(RemappingTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
-        set_target_properties(TypeLoweringTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
+        set_target_properties(StringTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
         set_target_properties(AttributePassthroughTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
         set_target_properties(MixedModeTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
         set_target_properties(NullableTests PROPERTIES COMPILE_FLAGS "-m32" LINK_FLAGS "-m32")
