+++ AdvancedDLSupport/Attributes/CallerFreeAttribute.cs
@@ -25,7 +25,7 @@ namespace AdvancedDLSupport
     /// <summary>
     /// Parameters which are marshalled into unmanaged memory with this attribute are freed by the caller after the conclusion of the call.
     /// </summary>
-    [PublicAPI, AttributeUsage(AttributeTargets.Parameter)]
+    [PublicAPI, AttributeUsage(AttributeTargets.Parameter | AttributeTargets.ReturnValue)]
     public class CallerFreeAttribute : Attribute
     {
     }
